refactor(bridge): isolate media download orchestration
